requirements.txt
albucore==0.0.13
albumentations==1.4.12
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
async-timeout==4.0.3
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
cookiecutter==2.6.0
easydict==1.13
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
fsspec==2024.6.1
gdown==5.2.0
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
locust==2.31.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numpy==1.26.2
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.18.1
# nvidia-nvjitlink-cu12==12.6.20
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==23.2
Pillow==10.0.1
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
pyzmq==26.1.0
qrcode==7.4.2
redis==5.0.8
repath==0.9.0
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
sympy==1.13.1
text-unidecode==1.3
tifffile==2024.7.24
timm==1.0.8
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.4
transparent-background==1.3.1
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wget==3.2
zope.event==5.0
zope.interface==7.0.1
